add WorkEmbed component
