+++ catalogue/webapp/pages/embed.js
@@ -1,10 +1,8 @@
 // @flow
-import {Fragment, Component} from 'react';
+import {Component} from 'react';
 import ReactGA from 'react-ga';
 import fetch from 'isomorphic-unfetch';
-import {iiifImageTemplate} from '@weco/common/utils/convert-image-uri';
-import ImageViewer2 from '@weco/common/views/components/ImageViewer/ImageViewer2';
-import Image from '@weco/common/views/components/Image/Image';
+import WorkEmbed from '@weco/common/views/components/WorkEmbed/WorkEmbed';
 
 // TODO: Find out where to get these types
 class Embed extends Component<{| work: Object, withViewer: boolean |}> {
@@ -33,43 +31,7 @@ class Embed extends Component<{| work: Object, withViewer: boolean |}> {
 
   render() {
     const {work, withViewer} = this.props;
-    const [iiifImageLocation] = work.items.map(
-      item => item.locations.find(
-        location => location.locationType === 'iiif-image'
-      )
-    );
-    const iiifInfoUrl = iiifImageLocation && iiifImageLocation.url;
-    const iiifImage = iiifImageTemplate(iiifInfoUrl);
-    const imageUrl = iiifImage({width: 800});
-
-    return (
-      <Fragment>
-        <div className='enhanced' style={{
-          maxHeight: '95vh',
-          maxWidth: '100vw',
-          textAlign: 'center',
-          position: 'relative'
-        }}>
-          <Fragment>
-            {withViewer && <ImageViewer2
-              contentUrl={imageUrl}
-              id={work.id}
-              width={800}
-              trackTitle={work.title}
-            />}
-
-            {!withViewer && <Image
-              extraClasses='margin-v-auto width-auto inherit-max-height'
-              width={800}
-              defaultSize={800}
-              contentUrl={imageUrl}
-              lazyload={false}
-              sizesQueries='(min-width: 860px) 800px, calc(92.59vw + 22px)'
-              alt='' />}
-          </Fragment>
-        </div>
-      </Fragment>
-    );
+    return (<WorkEmbed work={work} withViewer={withViewer} />);
   }
 }
