+++ common/model/work.js
@@ -1,0 +1,5 @@
+// @flow
+
+// Until we find a way of autogenerating this, not sure if it makes sense to
+// just leave as an `Object`
+export type Work = Object;
